sync DB

+++ synchronizing.py
@@ -2,7 +2,7 @@ import mariadb
 import os
 from dotenv import load_dotenv
 
-class Connections():
+class Connections(): #connection DB, dataBasa = base EspoCRM, commandData = sql command, soed = for inseert, update and delete command
     def connect(self, dataBasa, commandData, soed):
         try:
             dotenv_path = os.path.join(os.path.dirname(__file__), '.env')
@@ -17,40 +17,41 @@ class Connections():
             zapros = (self.conn_params)
             connection = mariadb.connect(**zapros)
             cursor = connection.cursor()
-            # print(commandData)
-            # print(soed)
             if soed == False:
                 cursor.execute(f"""{commandData}""")
                 row = cursor.fetchall()
                 cursor.close()
                 connection.close()
-                # print(row)
                 return row
             if soed:
                 cursor.execute(f"""{commandData}""", soed)
                 connection.commit()
                 cursor.close()
                 connection.close()
-                # print(soed)
                 return True
         except:
-            print('error')
+            print('error connect')
 
-class Data(Connections):
-    def sravnenie(self, row, basa, database):
+class Data(Connections): # sync ExpoCRM-DB and python-DB
+    def sravnenie(self, row, basa, database): #
         result1 = list(set(basa) - set(row))
         result2 = list(set(row) - set(basa))
-        for i in result1:
-            commandi = (f"INSERT INTO notifications(notification_id, data, class, user) VALUES (?,?,?,?)")
-            self.connect(database, commandi, i)
-        for i in result2:
-            commandi = (f"DELETE FROM notifications WHERE notification_id=?")
-            turle = (i[0],)
-            self.connect(database, commandi, turle)
-        print(result1)
-        print(result2)
-
-    def addData(self, basa, database, commandMysql, soed):
+        if result1:
+            for i in result1:
+                commandi = (f"INSERT INTO notifications(notification_id, data, class, user) VALUES (?,?,?,?)")
+                self.connect(database, commandi, i)
+        if result2:
+            for i in result2:
+                commandi = (f"DELETE FROM notifications WHERE notification_id=?")
+                turle = (i[0],)
+                self.connect(database, commandi, turle)
+    def requestEspoCRM(self, database, commandMysql, soed):
+        try:
+            row = self.connect(database, commandMysql, soed)
+            return row
+        except:
+            print('error')
+    def addData(self, basa, database, commandMysql, soed): #connect pythhon-DB and request all reminder
         try:
             row = self.connect(database, commandMysql, soed)
             self.sravnenie(row, basa, database)
@@ -67,7 +68,6 @@ mySqlCommandSozdanie = """
             SELECT notification_id, data, class, user
             FROM notifications"""
 
-zapros = Connections()
-basa = zapros.connect('database', mySqlCommandProverka, False)
 zapros = Data()
+basa = zapros.requestEspoCRM('databaseOne', mySqlCommandProverka, False)
 zapros.addData(basa, 'databaseTwo', mySqlCommandSozdanie, False)
